# packages/openapi/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/v1/todos/{id}": {
      "get": {
        "description": "Get todo by ID",
        "summary": "Get todo by ID",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "getTodoById",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "active",
                        "completed",
                        "archived"
                      ]
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "low",
                        "medium",
                        "high"
                      ]
                    },
                    "dueDate": {
                      "type": "string",
                      "nullable": true
                    },
                    "completedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "parentTodoId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "categoryId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reminder": {
                          "type": "string"
                        },
                        "color": {
                          "type": "string"
                        },
                        "difficulty": {
                          "type": "number"
                        }
                      },
                      "nullable": true
                    },
                    "sortOrder": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "color": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "name",
                        "color",
                        "description",
                        "createdAt",
                        "updatedAt"
                      ],
                      "nullable": true
                    },
                    "children": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "draft",
                              "active",
                              "completed",
                              "archived"
                            ]
                          },
                          "priority": {
                            "type": "string",
                            "enum": [
                              "low",
                              "medium",
                              "high"
                            ]
                          },
                          "dueDate": {
                            "type": "string",
                            "nullable": true
                          },
                          "completedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "parentTodoId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "categoryId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "metadata": {
                            "type": "object",
                            "properties": {
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "reminder": {
                                "type": "string"
                              },
                              "color": {
                                "type": "string"
                              },
                              "difficulty": {
                                "type": "number"
                              }
                            },
                            "nullable": true
                          },
                          "sortOrder": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "title",
                          "description",
                          "status",
                          "priority",
                          "dueDate",
                          "completedAt",
                          "parentTodoId",
                          "categoryId",
                          "metadata",
                          "sortOrder",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "todoId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "content": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "todoId",
                          "userId",
                          "content",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "attachments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "todoId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "uploadedBy": {
                            "type": "string"
                          },
                          "downloadKey": {
                            "type": "string"
                          },
                          "fileSize": {
                            "type": "number",
                            "nullable": true
                          },
                          "mimeType": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "todoId",
                          "name",
                          "uploadedBy",
                          "downloadKey",
                          "fileSize",
                          "mimeType",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "title",
                    "description",
                    "status",
                    "priority",
                    "dueDate",
                    "completedAt",
                    "parentTodoId",
                    "categoryId",
                    "metadata",
                    "sortOrder",
                    "createdAt",
                    "updatedAt",
                    "category",
                    "children",
                    "comments",
                    "attachments"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "description": "Update todo",
        "summary": "Update todo",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "updateTodo",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "active",
                      "completed",
                      "archived"
                    ]
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high"
                    ]
                  },
                  "dueDate": {
                    "type": "string",
                    "nullable": true
                  },
                  "parentTodoId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "metadata": {
                    "type": "object",
                    "properties": {
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "reminder": {
                        "type": "string"
                      },
                      "color": {
                        "type": "string"
                      },
                      "difficulty": {
                        "type": "number"
                      }
                    },
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "active",
                        "completed",
                        "archived"
                      ]
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "low",
                        "medium",
                        "high"
                      ]
                    },
                    "dueDate": {
                      "type": "string",
                      "nullable": true
                    },
                    "completedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "parentTodoId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "categoryId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reminder": {
                          "type": "string"
                        },
                        "color": {
                          "type": "string"
                        },
                        "difficulty": {
                          "type": "number"
                        }
                      },
                      "nullable": true
                    },
                    "sortOrder": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "title",
                    "description",
                    "status",
                    "priority",
                    "dueDate",
                    "completedAt",
                    "parentTodoId",
                    "categoryId",
                    "metadata",
                    "sortOrder",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "description": "Delete todo",
        "summary": "Delete todo",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "deleteTodo",
        "responses": {
          "204": {
            "description": "204",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/todos/stats": {
      "get": {
        "description": "Get todo statistics",
        "summary": "Get todo statistics",
        "tags": [
          "Todo"
        ],
        "parameters": [],
        "operationId": "getTodoStats",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "draft": {
                      "type": "number"
                    },
                    "active": {
                      "type": "number"
                    },
                    "completed": {
                      "type": "number"
                    },
                    "archived": {
                      "type": "number"
                    },
                    "overdue": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "total",
                    "draft",
                    "active",
                    "completed",
                    "archived",
                    "overdue"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/todos/{id}/attachments": {
      "post": {
        "description": "Upload a file attachment to a todo",
        "summary": "Upload attachment to todo",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "uploadTodoAttachment",
        "requestBody": {
          "description": "Body",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "todoId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "uploadedBy": {
                      "type": "string"
                    },
                    "downloadKey": {
                      "type": "string"
                    },
                    "fileSize": {
                      "type": "number",
                      "nullable": true
                    },
                    "mimeType": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "todoId",
                    "name",
                    "uploadedBy",
                    "downloadKey",
                    "fileSize",
                    "mimeType",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/todos/{id}/attachments/{attachmentId}": {
      "delete": {
        "description": "Delete a file attachment from a todo",
        "summary": "Delete todo attachment",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "deleteTodoAttachment",
        "responses": {
          "204": {
            "description": "204",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/todos/{id}/attachments/{attachmentId}/download": {
      "get": {
        "description": "Get a presigned URL to download an attachment",
        "summary": "Get attachment download URL",
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "getAttachmentPresignedURL",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": [
                    "url"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "info": {
    "version": "1.0.0",
    "title": "ExecuTask REST API - Documentation",
    "description": "ExecuTask REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}